moved user input event code into file DX11RendererNode_reactive.cs with partial class.
