+++ Nodes/VVVV.DX11.Nodes/Nodes/Renderers/Graphics/DX11RendererNode.cs
@@ -30,14 +30,11 @@ using FeralTic.DX11;
 using FeralTic.DX11.Resources;
 using VVVV.DX11.Nodes.Renderers.Graphics.Touch;
 using VVVV.DX11.Windows;
-using System.Reflection;
-using FormsMouseEventArgs = System.Windows.Forms.MouseEventArgs;
-using System.Reactive.Linq;
 
 namespace VVVV.DX11.Nodes
 {
-    [PluginInfo(Name="Renderer",Category="DX11",Author="vux,tonfilm",AutoEvaluate=true,
-        InitialWindowHeight=300,InitialWindowWidth=400,InitialBoxWidth=400,InitialBoxHeight=300, InitialComponentMode=TComponentMode.InAWindow)]
+    [PluginInfo(Name = "Renderer", Category = "DX11", Author = "vux,tonfilm", AutoEvaluate = true,
+        InitialWindowHeight = 300, InitialWindowWidth = 400, InitialBoxWidth = 400, InitialBoxHeight = 300, InitialComponentMode = TComponentMode.InAWindow)]
     public partial class DX11RendererNode : IPluginEvaluate, IDisposable, IDX11RendererHost, IDX11RenderWindow, IDX11Queryable, IUserInputWindow, IBackgroundColor, IPartImportsSatisfiedNotification
     {
         #region Touch Stuff
@@ -90,13 +87,14 @@ namespace VVVV.DX11.Nodes
                     handled = false;
                     break;
             }
-            base.WndProc(ref m);  // Call parent WndProc for default message processing.
 
             if (this.FInTouchUnhandle[0] == false)
             {
                 if (handled) // Acknowledge event if handled.
                     m.Result = new System.IntPtr(1);
             }
+
+            base.WndProc(ref m);  // Call parent WndProc for default message processing.
         }
 
         readonly int touchInputByteSize = Marshal.SizeOf(typeof(TOUCHINPUT));
@@ -175,31 +173,31 @@ namespace VVVV.DX11.Nodes
         [Import()]
         protected ILogger logger;
 
-        [Input("Layers", Order=1)]
+        [Input("Layers", Order = 1)]
         protected Pin<DX11Resource<DX11Layer>> FInLayer;
 
-        [Input("Clear",DefaultValue=1,Order = 2)]
+        [Input("Clear", DefaultValue = 1, Order = 2)]
         protected ISpread<bool> FInClear;
 
         [Input("Clear Depth", DefaultValue = 1, Order = 2)]
         protected ISpread<bool> FInClearDepth;
 
-        [Input("Background Color",DefaultColor=new double[] { 0,0,0,1 },Order=3)]
+        [Input("Background Color", DefaultColor = new double[] { 0, 0, 0, 1 }, Order = 3)]
         protected ISpread<Color4> FInBgColor;
 
-        [Input("VSync",Visibility=PinVisibility.OnlyInspector, IsSingle=true)]
+        [Input("VSync", Visibility = PinVisibility.OnlyInspector, IsSingle = true)]
         protected ISpread<bool> FInVsync;
 
-        [Input("Buffer Count", Visibility = PinVisibility.OnlyInspector, DefaultValue=1, IsSingle=true)]
+        [Input("Buffer Count", Visibility = PinVisibility.OnlyInspector, DefaultValue = 1, IsSingle = true)]
         protected IDiffSpread<int> FInBufferCount;
 
-        [Input("Do Not Wait", Visibility = PinVisibility.OnlyInspector, IsSingle=true)]
+        [Input("Do Not Wait", Visibility = PinVisibility.OnlyInspector, IsSingle = true)]
         protected ISpread<bool> FInDNW;
 
         [Input("Show Cursor", DefaultValue = 0, Visibility = PinVisibility.OnlyInspector)]
         protected IDiffSpread<bool> FInShowCursor;
 
-        [Input("Disable Shortcuts", DefaultValue = 0,IsSingle =true, Visibility = PinVisibility.OnlyInspector)]
+        [Input("Disable Shortcuts", DefaultValue = 0, IsSingle = true, Visibility = PinVisibility.OnlyInspector)]
         protected IDiffSpread<bool> FInDisableShortCuts;
 
         [Input("Refresh Rate", DefaultValue = 60, Visibility = PinVisibility.OnlyInspector)]
@@ -211,13 +209,13 @@ namespace VVVV.DX11.Nodes
         [Input("Fullscreen", Order = 5)]
         protected IDiffSpread<bool> FInFullScreen;
 
-        [Input("Enable Depth Buffer", Order = 6,DefaultValue=1)]
+        [Input("Enable Depth Buffer", Order = 6, DefaultValue = 1)]
         protected IDiffSpread<bool> FInDepthBuffer;
 
         [Input("Clear Depth Value", Order = 9, DefaultValue = 1)]
         protected ISpread<float> FInClearDepthValue;
 
-        [Input("AA Samples per Pixel", DefaultEnumEntry="1",EnumName="DX11_AASamples")]
+        [Input("AA Samples per Pixel", DefaultEnumEntry = "1", EnumName = "DX11_AASamples")]
         protected IDiffSpread<EnumEntry> FInAASamplesPerPixel;
 
         [Input("Enabled", DefaultValue = 1, Order = 9)]
@@ -229,7 +227,7 @@ namespace VVVV.DX11.Nodes
         [Input("Projection", Order = 11)]
         protected IDiffSpread<Matrix> FInProjection;
 
-        [Input("Aspect Ratio", Order = 12,Visibility=PinVisibility.Hidden)]
+        [Input("Aspect Ratio", Order = 12, Visibility = PinVisibility.Hidden)]
         protected IDiffSpread<Matrix> FInAspect;
 
         [Input("Crop", Order = 13, Visibility = PinVisibility.OnlyInspector)]
@@ -244,13 +242,13 @@ namespace VVVV.DX11.Nodes
         #endregion
 
         #region Output Pins
-        [Output("Mouse State",AllowFeedback=true)]
+        [Output("Mouse State", AllowFeedback = true)]
         protected ISpread<MouseState> FOutMouseState;
 
         [Output("Keyboard State", AllowFeedback = true)]
         protected ISpread<KeyboardState> FOutKState;
 
-        [Output("Touch Supported",IsSingle=true)]
+        [Output("Touch Supported", IsSingle = true)]
         protected ISpread<bool> FOutTouchSupport;
 
         [Output("Touch Data", AllowFeedback = true)]
@@ -264,13 +262,13 @@ namespace VVVV.DX11.Nodes
 
         protected ISpread<DX11Resource<DX11SwapChain>> FOuFS;
 
-        [Output("Present Time",IsSingle=true)]
+        [Output("Present Time", IsSingle = true)]
         protected ISpread<double> FOutPresent;
 
         [Output("Query", Order = 200, IsSingle = true)]
         protected ISpread<IDX11Queryable> FOutQueryable;
 
-        [Output("Control", Order = 201, IsSingle = true, Visibility = PinVisibility.OnlyInspector, AllowFeedback =true)]
+        [Output("Control", Order = 201, IsSingle = true, Visibility = PinVisibility.OnlyInspector, AllowFeedback = true)]
         protected ISpread<Control> FOutCtrl;
 
         [Output("Node Ref", Order = 201, IsSingle = true, Visibility = PinVisibility.OnlyInspector)]
@@ -321,12 +319,12 @@ namespace VVVV.DX11.Nodes
             if (!this.depthmanager.FormatChanged) // do not clear reset if format changed
             {
                 this.depthmanager.NeedReset = false;
-            } 
+            }
             else
             {
                 this.depthmanager.FormatChanged = false; //Clear flag ok
             }
-            
+
             if (FInAASamplesPerPixel.IsChanged || this.FInBufferCount.IsChanged || this.FInFlipSequential.IsChanged || this.FInRefreshRate.IsChanged)
             {
                 this.depthmanager.NeedReset = true;
@@ -360,9 +358,9 @@ namespace VVVV.DX11.Nodes
             }
 
             this.FOutKState[0] = new KeyboardState(this.FKeys);
-            this.FOutMouseState[0] = MouseState.Create(this.FMousePos.x, this.FMousePos.y, this.FMouseButtons.x > 0.5f, this.FMouseButtons.y > 0.5f, this.FMouseButtons.z> 0.5f, false, false, this.wheel);
+            this.FOutMouseState[0] = MouseState.Create(this.FMousePos.x, this.FMousePos.y, this.FMouseButtons.x > 0.5f, this.FMouseButtons.y > 0.5f, this.FMouseButtons.z > 0.5f, false, false, this.wheel);
             this.FOutBackBufferSize[0] = new Vector2D(this.Width, this.Height);
-            
+
             this.FOutTouchSupport[0] = this.touchsupport;
 
             this.FOutTouchData.SliceCount = this.touches.Count;
@@ -403,7 +401,7 @@ namespace VVVV.DX11.Nodes
         public void Render(DX11RenderContext context)
         {
             Device device = context.Device;
-            
+
             if (!this.updateddevices.Contains(context)) { this.Update(context); }
 
             if (this.rendereddevices.Contains(context)) { return; }
@@ -463,7 +461,7 @@ namespace VVVV.DX11.Nodes
                     {
                         this.EndQuery(context);
                     }
-                } 
+                }
                 catch (Exception ex)
                 {
                     exception = ex;
@@ -484,7 +482,7 @@ namespace VVVV.DX11.Nodes
         }
         #endregion
 
-        private void RenderSlice(DX11RenderContext context,DX11RenderSettings settings, int i, bool viewportpop)
+        private void RenderSlice(DX11RenderContext context, DX11RenderSettings settings, int i, bool viewportpop)
         {
             float cw = (float)this.ClientSize.Width;
             float ch = (float)this.ClientSize.Height;
@@ -568,9 +566,9 @@ namespace VVVV.DX11.Nodes
 
                 this.FInvalidateSwapChain = false;
 
-                #if DEBUG
+#if DEBUG
                 this.FOutBackBuffer[0][context].Resource.DebugName = "BackBuffer";
-                #endif
+#endif
                 this.depthmanager.NeedReset = true;
             }
 
@@ -604,16 +602,16 @@ namespace VVVV.DX11.Nodes
                 PresentFlags flags = this.FInDNW[0] ? (PresentFlags)8 : PresentFlags.None;
                 if (this.FInVsync[0])
                 {
-                    this.FOutBackBuffer[0][this.RenderContext].Present(1, flags); 
+                    this.FOutBackBuffer[0][this.RenderContext].Present(1, flags);
                 }
                 else
                 {
-                    this.FOutBackBuffer[0][this.RenderContext].Present(0, flags); 
+                    this.FOutBackBuffer[0][this.RenderContext].Present(0, flags);
                 }
             }
             catch
             {
-                
+
             }
 
             sw.Stop();
@@ -637,7 +635,7 @@ namespace VVVV.DX11.Nodes
 
         public IntPtr WindowHandle
         {
-            get 
+            get
             {
                 return this.Handle;
             }
@@ -692,253 +690,7 @@ namespace VVVV.DX11.Nodes
             this.FOutCtrl[0] = this;
             this.FOutRef[0] = (INode)this.FHost;
 
-            //add mouse event output for newer vvvv versions
-            var mouseType = typeof(Mouse);
-
-            //older mouse type
-            var olderParams = new Type[] { typeof(IObservable<MouseNotification>), typeof(bool) };
-            var olderConstructor = mouseType.GetConstructor(olderParams);
-            if(olderConstructor != null)
-            {
-                var oa = new OutputAttribute("Mouse Events");
-                oa.IsSingle = true;
-                var mouseEventOut = FIOFactory.CreatePin<Mouse>(oa);
-                mouseEventOut[0] = InitMouseEvents(olderConstructor, true);
-
-                CreateKeyboardEventsOut();
-                CreateTouchEventsOut();
-            }
-            else //newer mouse type
-            {
-                var newerParams = new Type[] { typeof(IObservable<MouseNotification>), typeof(bool), typeof(bool) };
-                var newerConstructor = mouseType.GetConstructor(newerParams);
-
-                if (newerConstructor != null)
-                {
-                    var oa = new OutputAttribute("Mouse Events");
-                    oa.IsSingle = true;
-                    var mouseEventOut = FIOFactory.CreatePin<Mouse>(oa);
-                    mouseEventOut[0] = InitMouseEvents(newerConstructor);
-
-                    CreateKeyboardEventsOut();
-                    CreateTouchEventsOut();
-                }
-            }
-        }
-
-        private Mouse InitMouseEvents(ConstructorInfo constructor, bool isOlderMouse = false)
-        {
-            var mouseDowns = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseDown")
-                .Select(p => p.EventArgs.ToMouseDownNotification(this));
-            var mouseMoves = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseMove")
-                .Select(p => p.EventArgs.ToMouseMoveNotification(this));
-            var mouseUps = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseUp")
-                .Select(p => p.EventArgs.ToMouseUpNotification(this));
-            var mouseClicks = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseClick")
-                .Select(p => p.EventArgs.ToMouseClickNotification(this));
-            var mouseDoubleClicks = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseDoubleClick")
-                .Select(p => p.EventArgs.ToMouseDoubleClickNotification(this));
-            var mouseWheels = Observable.FromEventPattern<FormsMouseEventArgs>(this, "MouseWheel")
-                .Select(p => p.EventArgs.ToMouseWheelNotification(this));
-
-            var mouseStream = mouseDowns
-                        .Merge<MouseNotification>(mouseMoves)
-                        .Merge(mouseUps)
-                        .Merge(mouseClicks)
-                        .Merge(mouseDoubleClicks)
-                        .Merge(mouseWheels);
-
-            if (isOlderMouse)
-                return (Mouse)constructor.Invoke(new object[] { mouseStream, false });
-            else
-                return (Mouse)constructor.Invoke(new object[] { mouseStream, false, true });
-        }
-
-        private void CreateKeyboardEventsOut()
-        {
-            var oa = new OutputAttribute("Keyboard Events");
-            oa.IsSingle = true;
-            var keyboardEventOut = FIOFactory.CreatePin<Keyboard>(oa);
-
-            var keyDowns = Observable.FromEventPattern<KeyEventArgs>(this, "KeyDown")
-                .Select(p => p.EventArgs.ToKeyDownNotification());
-            var keyUps = Observable.FromEventPattern<KeyEventArgs>(this, "KeyUp")
-                .Select(p => p.EventArgs.ToKeyUpNotification());
-            var keyPresses = Observable.FromEventPattern<KeyPressEventArgs>(this, "KeyPress")
-                .Select(p => p.EventArgs.ToKeyPressNotification());
-
-            var keyboardStream = keyDowns
-                .Merge<KeyNotification>(keyUps)
-                .Merge(keyPresses);
-
-            keyboardEventOut[0] = new Keyboard(keyboardStream);
-        }
-
-        private void CreateTouchEventsOut()
-        {
-            var oa = new OutputAttribute("Touch Events");
-            oa.IsSingle = true;
-            var touchEventOut = FIOFactory.CreatePin<TouchDevice>(oa);
-
-            var touchDowns = Observable.FromEventPattern<WMTouchEventArgs>(this, "TouchDown")
-                .Select(p => p.EventArgs.ToTouchDownNotification(this));
-            var touchMoves = Observable.FromEventPattern<WMTouchEventArgs>(this, "TouchMove")
-                .Select(p => p.EventArgs.ToTouchMoveNotification(this));
-            var touchUps = Observable.FromEventPattern<WMTouchEventArgs>(this, "TouchUp")
-                .Select(p => p.EventArgs.ToTouchUpNotification(this));
-
-            var touchStream = touchDowns
-                .Merge<TouchNotification>(touchMoves)
-                .Merge(touchUps);
-
-            touchEventOut[0] = new TouchDevice(touchStream);
-        }
-    }
-
-    public static class EventExtensions
-    {
-        //mouse
-        public static MouseMoveNotification ToMouseMoveNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseMoveNotification(args.Location, relativeTo.ClientSize);
-        }
-
-        public static MouseDownNotification ToMouseDownNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseDownNotification(args.Location, relativeTo.ClientSize, args.Button);
-        }
-
-        public static MouseUpNotification ToMouseUpNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseUpNotification(args.Location, relativeTo.ClientSize, args.Button);
-        }
-
-        public static MouseClickNotification ToMouseClickNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseClickNotification(args.Location, relativeTo.ClientSize, args.Button, 1);
-        }
-
-        public static MouseClickNotification ToMouseDoubleClickNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseClickNotification(args.Location, relativeTo.ClientSize, args.Button, 2);
-        }
-
-        public static MouseWheelNotification ToMouseWheelNotification(this FormsMouseEventArgs args, Control relativeTo)
-        {
-            return new MouseWheelNotification(args.Location, relativeTo.ClientSize, args.Delta);
-        }
-
-        //keyboard
-        public static KeyDownNotification ToKeyDownNotification(this KeyEventArgs eventArgs)
-        {
-            return new KeyDownNotification(eventArgs.KeyCode);
-        }
-
-        public static KeyUpNotification ToKeyUpNotification(this KeyEventArgs eventArgs)
-        {
-            return new KeyUpNotification(eventArgs.KeyCode);
-        }
-
-        public static KeyPressNotification ToKeyPressNotification(this KeyPressEventArgs eventArgs)
-        {
-            return new KeyPressNotification(eventArgs.KeyChar);
-        }
-
-        //touch
-
-        enum TouchNotificationAge { Older, Intermediate, Newer};
-        static TouchNotificationAge TouchAge;
-
-        //cash found constructors as this gets called for each event
-        static ConstructorInfo STouchDownNotificationConstructor;
-        static ConstructorInfo STouchMoveNotificationConstructor;
-        static ConstructorInfo STouchUpNotificationConstructor;
-
-        public static TouchDownNotification ToTouchDownNotification(this WMTouchEventArgs eventArgs, Control relativeTo)
-        {
-            if (STouchDownNotificationConstructor == null) //find correct constructor
-                STouchDownNotificationConstructor = FindTouchNotificationConstructor<TouchDownNotification>();
-
-            return ConstructTouchNotification<TouchDownNotification>(STouchDownNotificationConstructor, eventArgs, relativeTo);
-        }
-
-
-        public static TouchMoveNotification ToTouchMoveNotification(this WMTouchEventArgs eventArgs, Control relativeTo)
-        {
-            if (STouchMoveNotificationConstructor == null) //find correct constructor
-                STouchMoveNotificationConstructor = FindTouchNotificationConstructor<TouchMoveNotification>();
-
-            return ConstructTouchNotification<TouchMoveNotification>(STouchMoveNotificationConstructor, eventArgs, relativeTo);
-
-        }
-
-        public static TouchUpNotification ToTouchUpNotification(this WMTouchEventArgs eventArgs, Control relativeTo)
-        {
-            if (STouchUpNotificationConstructor == null) //find correct constructor
-                STouchUpNotificationConstructor = FindTouchNotificationConstructor<TouchUpNotification>();
-
-            return ConstructTouchNotification<TouchUpNotification>(STouchUpNotificationConstructor, eventArgs, relativeTo);
-        }
-
-        static T ConstructTouchNotification<T>(ConstructorInfo constructor, WMTouchEventArgs eventArgs, Control relativeTo) where T : TouchNotification
-        {
-            switch (TouchAge)
-            {
-                case TouchNotificationAge.Newer:
-                    return (T)constructor.Invoke(new object[] { eventArgs.Location(), relativeTo.ClientSize, eventArgs.Id, eventArgs.IsPrimaryContact, eventArgs.ContactArea(), eventArgs.TouchDeviceID });
-
-                case TouchNotificationAge.Intermediate:
-                    return (T)constructor.Invoke(new object[] { eventArgs.Location(), relativeTo.ClientSize, eventArgs.Id, eventArgs.IsPrimaryContact, eventArgs.ContactArea() });
-
-                case TouchNotificationAge.Older:
-                    return (T)constructor.Invoke(new object[] { eventArgs.Location(), relativeTo.ClientSize, eventArgs.Id, eventArgs.ContactArea() });
-            }
-
-            return null;
-        }
-
-        static ConstructorInfo FindTouchNotificationConstructor<T>() where T : TouchNotification
-        {
-            var touchNotificationType = typeof(T);
-
-            var newerParams = new Type[] { typeof(Point), typeof(Size), typeof(int), typeof(bool), typeof(Size), typeof(long) };
-            var newerConstructor = touchNotificationType.GetConstructor(newerParams);
-
-            if (newerConstructor != null)
-            {
-                TouchAge = TouchNotificationAge.Newer;
-                return newerConstructor;
-            }
-
-            var intermediateParams = new Type[] { typeof(Point), typeof(Size), typeof(int), typeof(bool), typeof(Size) };
-            var intermediateConstructor = touchNotificationType.GetConstructor(intermediateParams);
-
-            if (intermediateConstructor != null)
-            {
-                TouchAge = TouchNotificationAge.Intermediate;
-                return STouchUpNotificationConstructor = intermediateConstructor;
-            }
-
-            var olderParams = new Type[] { typeof(Point), typeof(Size), typeof(int), typeof(Size) };
-            var olderConstructor = touchNotificationType.GetConstructor(olderParams);
-
-            if (olderConstructor != null)
-            {
-                TouchAge = TouchNotificationAge.Older;
-                return olderConstructor;
-            }
-
-            return null;
-        }
-
-        public static Point Location(this WMTouchEventArgs args)
-        {
-            return new Point(args.LocationX, args.LocationY);
-        }
-
-        public static Size ContactArea(this WMTouchEventArgs args)
-        {
-            return new Size(args.ContactX, args.ContactY);
+            CreateUserInputEventPins();
         }
     }
 }
